requirements.txt
Jinja2==3.1.2
nbconvert==7.2.1
requests==2.32.3
packaging==21.3
xlrd==2.0.1
openpyxl==3.2.0b1
lxml==5.3.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
setuptools==58.0.4
future==0.18.2
docx2pdf==0.1.8
python-docx==1.1.2
fpdf==1.7.2
fpdf2==2.8.2
plotly==5.24.1
numpy>=1.21.0
nltk==3.9.1  